Refactor hatch code into Hatch.h and Hatch.cpp
